requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
scikit-learn>=1.0.0
pandas>=1.3.0
seaborn>=0.11.0
numba>=0.56.0
psutil>=5.8.0
